requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiosignal
alembic
annotated-types
anyio
asttokens
attrs
azure-core
azure-storage-blob
blinker
cachetools
certifi
cffi
charset-normalizer
click
cloudpickle
colorama
comm
contourpy
cryptography
cycler
databricks-sdk
datasets
debugpy
decorator
Deprecated
dill
docker
executing
fastapi
filelock
Flask
fonttools
frozenlist
fsspec
gitdb
GitPython
google-auth
graphene
graphql-core
graphql-relay
greenlet
h11
huggingface-hub
idna
importlib_metadata
ipykernel
ipython
ipython_pygments_lexers
isodate
itsdangerous
jedi
Jinja2
joblib
jupyter_client
jupyter_core
kiwisolver
Mako
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mlflow
mlflow-skinny
mpmath
multidict
multiprocess
nest-asyncio
networkx
ngrok
numpy
opentelemetry-api
opentelemetry-sdk
opentelemetry-semantic-conventions
packaging
pandas
parso
peft
pillow
platformdirs
prompt_toolkit
propcache
protobuf
psutil
pure_eval
py4j
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
Pygments
pyparsing
pyspark
python-dateutil
python-dotenv
pytz
PyYAML
pyzmq
regex
requests
rsa
safetensors
scikit-learn
scipy
sentencepiece
six
smmap
sniffio
SQLAlchemy
sqlparse
stack-data
starlette
sympy
threadpoolctl
tokenizers
torch
tornado
tqdm
traitlets
transformers
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
waitress
wcwidth
Werkzeug
wrapt
xxhash
yarl
zipp
